refactor lots of code out into helpers.scala
